Add a section of handling still-too-big wasm
Comparisons of WebAssembly in terms of code size tend to only take a
look at the size of the wasm file itself and stop there. In terms of
benchmarking though code size is rarely the only factor to consider!
This commit adds a section to the code size section of the book to list
a few other factors which are important to consider when profiling and
measuring code size of wasm files.